+++ src/pages/Notice/Search/Components/NotificationBadge/NotificationBadge.module.css
