+++ src/pages/Notice/Search/Components/NotificationBadge/NotificationBadge.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
 import styles from "./NotificationBadge.module.css";
-import notificationIcon from "../../../../assets/icon/notice/search/bell.svg";
+import notificationIcon from "../../../../../assets/icon/notice/search/bell.svg";
 
 interface NotificationBadgeProps {
   keyword: string;
