+++ src/pages/Notice/Search/Components/SearchHistory/SearchHistory.module.css
