+++ src/pages/Notice/Search/Components/SearchHistory/SearchHistory.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import SearchTag from "./SearchTag";
-import SearchHistoryMenu from "./SearchHistoryMenu";
+import SearchTag from "../SearchTag/SearchTag";
+import SearchHistoryMenu from "../SearchHistoryMenu/SearchHistoryMenu";
 import styles from "./SearchHistory.module.css";
-import kebabIcon from "../../../../assets/icon/notice/search/kebab.svg";
+import kebabIcon from "../../../../../assets/icon/notice/search/kebab.svg";
 
 interface SearchHistoryProps {
   searchTerms: string[];
