+++ src/pages/Notice/Search/Components/SearchResult/SearchResult.module.css
