+++ src/pages/Notice/Search/Components/SearchResult/SearchResult.tsx
@@ -1,8 +1,8 @@
 import type React from "react";
 import styles from "./SearchResult.module.css";
-import type { NoticeItem } from "../../../../services/NoticeService";
-import NoticeList from "../../NoticeList";
-import searchresultIcon from "../../../../assets/icon/notice/search/searchresult.svg";
+import type { NoticeItem } from "../../../../../services/NoticeService";
+import NoticeList from "../../../Notice/NoticeList/NoticeList";
+import searchresultIcon from "../../../../../assets/icon/notice/search/searchresult.svg";
 
 interface SearchResultProps {
   filteredNotices: NoticeItem[];
