+++ src/pages/Notice/Search/Components/SearchTag/SearchTag.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
 import styles from "./SearchTag.module.css";
-import closeIcon from "../../../../assets/icon/notice/search/remove.svg";
+import closeIcon from "../../../../../assets/icon/notice/search/remove.svg";
 
 interface SearchTagProps {
   text: string;
