+++ src/pages/Notice/Search/Search.tsx
@@ -2,12 +2,12 @@ import type React from "react";
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import Header from "../../../components/Header/Header";
-import SearchInput from "./Components/SearchInput";
-import SearchHistory from "./Components/SearchHistory";
-import TagButtons from "./Components/TagButtons";
-import SearchNoticeList from "./Components/NoticeList";
-import SearchResult from "./Components/SearchResult";
-import NotificationBadge from "./Components/NotificationBadge";
+import SearchInput from "./Components/SearchInput/SearchInput";
+import SearchHistory from "./Components/SearchHistory/SearchHistory";
+import TagButtons from "./Components/TagButtons/TagButtons";
+import SearchNoticeList from "./Components/NoticeList/NoticeList";
+import SearchResult from "./Components/SearchResult/SearchResult";
+import NotificationBadge from "./Components/NotificationBadge/NotificationBadge";
 import { getAllNotices } from "../../../services/NoticeService";
 import type { NoticeItem } from "../../../services/NoticeService";
 import styles from "./Search.module.css";
